allow loosing footing and drowning also in shallow ocean water
The logic is that "shallow" ocean water is not really _that_ shallow
compared to normal shallow water - since the only other ocean water
is so deep that it's impassable, "shallow" ocean water is more likely
to be chest-deep than ankle-deep. And with waves, which make ocean
water to be sort of moving. Even realistically it should be rather unlikely
to find a larger area of really shallow ocean water where one wouldn't
be at any risk.