do not copy collections on every map tick
That's expensive to do often. Read-only access to collections should be
thread-safe, and this code does not run in parallel with other code that
would modify the collections. Also removed null checks as all those
collections are set in the constructor and should never be null.

The only code that needs explicit thread-safety is when accessed
from multiple threads and at least one of those accesses is write access.